Made changes to the backend to use the authentication protocols added previously. - Added a get_current_user dependency in auth.py allowing any route to fetch the complete user ID object from the db for valid and authenticated users. - Created a type alias for the above dependency to make it easier to use it across other modules. - Removed user_id parameter requirement from PostCreate schema in schemas.py file. This is done to capture the user ID from the token instead of the api request body so that only logged in users, who are authenticated,  are able to create posts. - Made changes to router/posts.py file to reflect the get_current_user dependency - 	- Modified the create_post endpoint to eliminiate checking for the user_id and instead using the CurrentUser dependency to verify user. 	- Modified the PUT, PATCH and DELETE endpoints to check authorized users. Again, remmoved extra lines code which was checking the existence of the user. - Made changes to routers/users.py file - 	- Modified the /me endpoint to user CurrentUser for getting authenticated users. 	- Modified the PATCH and DELETE endpoints to check for authorized users.

+++ routers/posts.py
@@ -9,6 +9,8 @@ import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
+from auth import CurrentUser
+
 router = APIRouter()
 
 
@@ -19,20 +21,14 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     posts = result.scalars().all()
     return posts
 
+# By adding the current_user dependency, this route is protected. If someone tries to call this endpoint without a valid token, they will get a 401 unauthorized before the fucntion even runs.
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found"
-        )
+async def create_post(post: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id
+        user_id=current_user.id # this comes from the authenticated token instead of the api request body
     )
 
     db.add(new_post)
@@ -52,28 +48,20 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]): 
 
 # this endpoint is for a PUT request which requires all the information to be provided for an update; Hence, we will be using PostCreate instead of PostUpdate because that already requires all the information.
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int, post_data: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 
-    # checking whether the new user_id sent by the PUT request exists; we will only do this if the client sent a user_id which is different from what was already present in the original post
-    # we only want to allow users who created a post to update or delete them
-    if post_data.user_id != post.user_id:
-        result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = result.scalars().first()
+    # 401 is unauthorized, that means you are not authenticated, i.e. if there is a missing or invalid token.
+    # 403 means you're authenticated but you don't have permission for this action.
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found"
-            )
-        
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -81,12 +69,15 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
 
 
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int, post_data: PostUpdate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
     # exclude_unset attribute only sets the values which were sent in the api request; it is essential because otherwise it will set all the values not sent in the request to default values
     update_data = post_data.model_dump(exclude_unset=True)
@@ -101,13 +92,15 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
 
 # For DELETE requests, we usually return a 204 No Content Response which means that the request succeeded but there is no response body. Hence we don't have a response_model here.
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT )
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]): # using type hinting helps FastAPI to automatically validate the input
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]): # using type hinting helps FastAPI to automatically validate the input
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
+
     await db.delete(post)
     await db.commit()
-# Later when we add authentication, we'll add ownership checks so that only the author of the post is allowed to delete it.
